requirements.txt
Django==3.0
djangorestframework==3.11.0
django-environ==0.4.5
psycopg2-binary==2.8.6
celery==5.1.2
redis==3.5.3
django-excel==0.0.10
pytest-django== 4.4.0

